Use character strings to designate permissions in set_permissions().
I borrowed this idea from Python's `open()` builtin. It's an easier syntax and
less lengthy than a complete, space-separated string or tuple of strings.

The change is actually too powerful (or maybe too dumb?) because it completely
ignores any characters that aren't specified in the permission set. This will be
fixed in a future commit that addresses input validation.